+++ src/Pgvector.EntityFrameworkCore/HalfvecTypeMappingSourcePlugin.cs
@@ -1,0 +1,11 @@
+using Microsoft.EntityFrameworkCore.Storage;
+
+namespace Pgvector.EntityFrameworkCore;
+
+public class HalfvecTypeMappingSourcePlugin : IRelationalTypeMappingSourcePlugin
+{
+    public RelationalTypeMapping? FindMapping(in RelationalTypeMappingInfo mappingInfo)
+        => mappingInfo.ClrType == typeof(HalfVector)
+            ? new HalfvecTypeMapping(mappingInfo.StoreTypeName ?? "halfvec")
+            : null;
+}