+++ src/Pgvector.EntityFrameworkCore/VectorDbFunctionsExtensions.cs
@@ -4,15 +4,15 @@ namespace Pgvector.EntityFrameworkCore;
 
 public static class VectorDbFunctionsExtensions
 {
-    public static double L2Distance(this Vector a, Vector b)
+    public static double L2Distance(this object a, object b)
         => throw new InvalidOperationException(CoreStrings.FunctionOnClient(nameof(L2Distance)));
 
-    public static double MaxInnerProduct(this Vector a, Vector b)
+    public static double MaxInnerProduct(this object a, object b)
         => throw new InvalidOperationException(CoreStrings.FunctionOnClient(nameof(MaxInnerProduct)));
 
-    public static double CosineDistance(this Vector a, Vector b)
+    public static double CosineDistance(this object a, object b)
         => throw new InvalidOperationException(CoreStrings.FunctionOnClient(nameof(CosineDistance)));
 
-    public static double L1Distance(this Vector a, Vector b)
+    public static double L1Distance(this object a, object b)
         => throw new InvalidOperationException(CoreStrings.FunctionOnClient(nameof(L1Distance)));
 }