+++ src/Pgvector/Npgsql/VectorExtensions.cs
@@ -8,6 +8,7 @@ public static class VectorExtensions
     public static INpgsqlTypeMapper UseVector(this INpgsqlTypeMapper mapper)
     {
         mapper.AddTypeInfoResolverFactory(new VectorTypeInfoResolverFactory());
+        mapper.AddTypeInfoResolverFactory(new HalfvecTypeInfoResolverFactory());
         return mapper;
     }
 }