+++ tests/Pgvector.CSharp.Tests/EntityFrameworkCoreTests.cs
@@ -36,6 +36,9 @@ public class Item
 
     [Column("embedding", TypeName = "vector(3)")]
     public Vector? Embedding { get; set; }
+
+    [Column("half_embedding", TypeName = "halfvec(3)")]
+    public HalfVector? HalfEmbedding { get; set; }
 }
 
 public class EntityFrameworkCoreTests
@@ -49,15 +52,16 @@ public class EntityFrameworkCoreTests
         var databaseCreator = ctx.GetService<IRelationalDatabaseCreator>();
         databaseCreator.CreateTables();
 
-        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 1 }) });
-        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 2, 2, 2 }) });
-        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 2 }) });
+        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 1 }), HalfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)1 }) });
+        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 2, 2, 2 }), HalfEmbedding = new HalfVector(new Half[] { (Half)2, (Half)2, (Half)2 }) });
+        ctx.Items.Add(new Item { Embedding = new Vector(new float[] { 1, 1, 2 }), HalfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)2 }) });
         ctx.SaveChanges();
 
         var embedding = new Vector(new float[] { 1, 1, 1 });
         var items = await ctx.Items.FromSql($"SELECT * FROM efcore_items ORDER BY embedding <-> {embedding} LIMIT 5").ToListAsync();
         Assert.Equal(new int[] { 1, 3, 2 }, items.Select(v => v.Id).ToArray());
         Assert.Equal(new float[] { 1, 1, 1 }, items[0].Embedding!.ToArray());
+        Assert.Equal(new Half[] { (Half)1, (Half)1, (Half)1 }, items[0].HalfEmbedding!.ToArray());
 
         items = await ctx.Items.OrderBy(x => x.Embedding!.L2Distance(embedding)).Take(5).ToListAsync();
         Assert.Equal(new int[] { 1, 3, 2 }, items.Select(v => v.Id).ToArray());
@@ -70,6 +74,10 @@ public class EntityFrameworkCoreTests
         Assert.Equal(3, items[2].Id);
 
         items = await ctx.Items.OrderBy(x => x.Embedding!.L1Distance(embedding)).Take(5).ToListAsync();
+        Assert.Equal(new int[] { 1, 3, 2 }, items.Select(v => v.Id).ToArray());
+
+        var halfEmbedding = new HalfVector(new Half[] { (Half)1, (Half)1, (Half)1 });
+        items = await ctx.Items.OrderBy(x => x.HalfEmbedding!.L2Distance(halfEmbedding)).Take(5).ToListAsync();
         Assert.Equal(new int[] { 1, 3, 2 }, items.Select(v => v.Id).ToArray());
 
         items = await ctx.Items